pyproject.toml
[project]
name = "lit-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "torch<2.5,>=2.4",
    "torchvision>=0.19.0",
    "safetensors>=0.4.5",
    "hf-transfer>=0.1.8",
    "wandb>=0.18.7",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.2",
    "lightning>=2.4.0",
    "cheesechaser>=0.2.0",
    "pydantic>=2.10.2",
    "transformers>=4.47.0",
    "webdataset>=0.2.100",
    "datasets>=3.2.0",
]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
    "ruff>=0.8.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "pytest>=8.3.4",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu124", marker = "platform_system != 'Darwin'" }]
torchvision = [
    { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
